improvement: improve flags handling
added a library for flags parsing that provides more advanced options.
removed '-so' and made it the default behavior. this is less
destructive and should provide more safety agaisnt accidents.
removed '-si' as this behaviour can be inferred from the absence of lua
scripts.
added '-i' option to specify the editing must be 'in-place' aka the user
commits to the changes made to the source.