feature: add spaces_before_call style option
This commit introduces a new feature to control the amount of spaces
between the a function identifier or expression, and the function
arguments if the paranthesis is ommited. Such omition is often used
in `require` or in a function style manner.

This commit also removes the shouldKeepSemicolon function, now all
semicolons are written to the file. I've made this change because
the heuristic is not configurable and it depended on assumptions that
could intervene with real-world code. Besides it broke the feature here
introduced, for some reason.

Signed-off-by: Pedro Tammela <pctammela@gmail.com>