Fix bug with option spaces-inside-functioncall-parens.
When the spaces_inside_functioncall_parens option is enabled,
there is an edge case that does not work.  That is, when the
line that is formatted without the option enabled fits on a
line, but then when formatted with the option enabled, the length
of the line gets pushed just over the column limit.  In that
case, the spaces are not inserted properly.  This should fix
that.

Added test case for this which fails without this fix.